Implemented MPRecyclerViewAdapter
