+++ features/mpaccount/src/main/java/com/jpp/mpaccount/account/lists/UserMoviesAdapter.kt
@@ -6,11 +6,10 @@ import androidx.databinding.DataBindingUtil
 import androidx.recyclerview.widget.RecyclerView
 import com.jpp.mpaccount.R
 import com.jpp.mpaccount.databinding.ListItemUserMovieBinding
+import com.jpp.mpdesign.adapters.MPRecyclerViewAdapter
 
 internal class UserMoviesAdapter(private val itemSelectionListener: (UserMovieItem) -> Unit) :
-    RecyclerView.Adapter<UserMoviesAdapter.ViewHolder>() {
-
-    private var itemList = ArrayList<UserMovieItem>()
+    MPRecyclerViewAdapter<UserMovieItem, UserMoviesAdapter.ViewHolder>() {
 
     override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): ViewHolder {
         return ViewHolder(
@@ -24,16 +23,7 @@ internal class UserMoviesAdapter(private val itemSelectionListener: (UserMovieIt
     }
 
     override fun onBindViewHolder(holder: ViewHolder, position: Int) {
-        holder.bindMovieItem(itemList[position], itemSelectionListener)
-    }
-
-    override fun getItemCount(): Int = itemList.size
-
-    fun submitList(newData: List<UserMovieItem>) {
-        val currentLastIndex = itemList.size
-        itemList.clear()
-        itemList.addAll(newData)
-        notifyItemRangeChanged(currentLastIndex, itemList.size)
+        holder.bindMovieItem(getItem(position), itemSelectionListener)
     }
 
     class ViewHolder(private val itemBinding: ListItemUserMovieBinding) :
